requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
groq==0.4.2
tavily==0.2.0
httpx==0.27.0
